chore: moved trpc utils to be under trpc folder
